Add interface console to play around
This commit adds an interactive console to test out the actual
API interactions. This commit also adds a `.sample.pryrc` file
and you can easily follow the instructions on the `#playbox`
section to set it up so you don't have to configure the client
every time you want to test it out.