Let's be declarative for class methods
In our API Resources, we are providing an easier interface to
call it's instance methods using the class level methods, and
we are using `method_missing` as fallback, and then it would
check if that method has been defined as public method or not
and if it did then it would create a new instance and invoke
that method, But that has some potential drawbacks.

So, after looking into that we only need `two/three` interface
so, let's be declarative about those and define those method
explicitly, it also ensures everything works as before.