Add interface to order a SSL Plus
This commit adds the interface to order a SSL Plus certificate.
This usages the ruby's hash arguments so it will do the basic
validation for the required keys. If any of the required keys
are missing then it will raise an error before making actual
API Request. Usages:

`Digicert::Order::SSLPlus.create(
  certificate: {
    common_name: "digicert.com",
    csr: "------ [CSR HERE] ------",
    signature_hash: "sha256",

    organization_units: ["Developer Operations"],
    server_platform: { id: 45 },
    profile_option: "some_ssl_profile",
  },

  organization: { id: 117483 },
  validity_years: 3,
  custom_expiration_date: "2017-05-18",
  comments: "Comments for the the approver",
  disable_renewal_notifications: false,
  renewal_of_order_id: 314152,
  payment_method: "balance",
)`