Add interface to list all templates
Container Templates define a set of features that are available to a
container. This commit add the interface to retrieve the list of all
available container templates. To retrieve the list you need to pass
the `parent_container_id` and it will return the objects

`Digicert::ContainerTemplate.all(container_id)`