Add interface to fetch a template details
This commit adds the interface to fetch a specific template details,
The way Digicert works is, we need provide the `template_id`, along
with `container_id` to retrieve the details for a template.

This commit adds the interface that override the default `fetch`
method signature, and now it expects two keys that is required to
specify the `container_id` and `template_id`, and as being a hash
arguments we don't need to provide those arguments in any specific
order. Usages

`Digicert::ContainerTemplate.fetch(
  container_id: container_id, template_id: template_id,
)`