Add interface to create a container
Container is an Operational Division used to model your organizational
structure. The features of the container you create are determined by
its Container Template. When you create a new container, you may also
specify the name, email address, and access role of a user for the
container.

This commit adds the interface to implement the above behavior and
it also expects one additional `container_id` along with the other
attributes.

`Digicert::Container.create(
  template_id: 5,
  container_id: 123_456_789,
  name: "History Department",
  description: "History, Civ, Ancient Languages",
)`