Add `#parse` to `Digicert::Request`
The existing interface for `Digicert::Request` might creates some
confusion, as we have a method name `#run` and that is actually
running the request and at the same time it is also parsing the
response, which is bit confusing.

This commit rename it to `#parse` and add one additional `#run`
method, which will only run the API request without doing any
parsing. So if we do not need to parse the response then we can
use the `#run` but if we do then we can use `#parse`.