Add interface to reissue an order
This commit adds an interface to reissue an existing order. By
default it usages the existing order details, but if the second
hash key `:certificate` is provided then it will use the updated
attributes and create a reissue request. Refer to the readme for
more details on that.

`Digicert::OrderReissuer.create(order_id: order_id)`