Remove unused codes from the gem
There is a significant amount of changes since we started the gem
development, we removed any third party gem dependencies, we wrote
our own simple client and we also did some changes on the existing
interfaces, but those codes are still there. This commit removes
those codes and updated those dependencies. Here is the highlight

* Remove third-party gem dependencies
* Replace `curb` with a simple net/http client
* Remove existing/unused codes form the gem