Remove `initialize` from subclasses
Currently, each of the `Digicert::Base` subclass was rewriting
the `initialize` method to extract some of the attributes that
is local to that class. This way every class know too much than
they actually need to and if they forget to call the `super`
than those won't work properly. Which is kind of against the
single responsibility principle and might be hard to maintain
in the long run.

This commit removes this by implementing hook kind of method,
so if we needs to extract some ids than we do not need to
rewrite the whole initialization method but all we need to do
is define `extract_local_attribute_ids` and extract the attrs.
This was every class does not need to know how to implement the
`initialize` that does not break anything it was not supposed
to do.