Do not expose `find` for every class
We have the `find` method defined in the `Base` class and it's
not exactly a inheritable behavior by all of it's subclasses,
so it makes more sense to extract it to a module, and add it
whenever it is necessary. This was we are not exposing it to
every single subclasses.