Remove multiple responsibility from base
Currently, the `Digicert::Base` is handling lots of responsibilities,
although all of them are not common across all of the sub classes. It
works for now but it might be hard to maintain and it's exposing the
interfaces that's might not work as expected.

This commit remove multiple responsibilities from `Digicert::Base`,
it only keeps the behavior that's common from any of it's subclasses,
and this also extract the behaviors to modules. Now we need to add
those module explicitly where we want to adds those functionality.

The by product of doing this is, it provides a cleaner interface,
and developer can instantly get the ideas just looking into those
classes.