Re-organize interface documentations
Currently, we don't have any specific documentation structure for
the content in the readme, which is hard to find an appropriate
interface.

This commit re-organize the documentation and structures it as
they did on the `Digicert` documentation, so this should make
it user to find any interface user want to use.