Change response type to be configurable
Currently we are using `Digicert::ResponseObject` class to parse
our existing API responses, which is basically a wrapper on top
of the `OpenStruct`. Normally it's slower then ruby's basic type
but the amount of data we would be dealing with is very small so
it should not create any issues.

But let's be good, let's allow the user to configure the response
type, so if they prefer then they can choose to parse their data
in ruby's fastest `hash` type.

`Digicert.configure do |config|
  config.response_type = "hash"
end`