Extract `Digicert::Request` reference to base
We extracted pretty much most of the actions in separated modules,
but there are still some use case scenario where we need to create
a new `Digicert::Request` instance and invoke the public API. But
this creates some sort of coupling and it might make it harder in
the long run.

This commit extract the `Digicert::Request` reference to the base
class and expose it as a `request_klass`, so if we need to change
then it would be the only place that will require changes.