Add instruction to generate the doc
Once we have made this gem public then we can use `Rubydoc.info` to
generate the documentation, but as long as we are for internal use
then we might need a better way to generate the updated doc.

This commit adds the documentation generation instructions, so our
developer can easily generate one whenever necessary.