Add container management request spec
As of now, our test suite is using an API response stubs, which
is identical to actual response but way much faster. But in some
time there might be some changes in the actual API, so it would
be best to have some specs that can test actual API response

This commit adds mechanism so we can add an `api_call: true` tag
to any specs and that will actually performs an API call.

We also changed our `spec_helper`, by default those slow, real
API call will be skipped, but we we can run those by running
`bin/rspec --tag api_call`