Add domain management request specs
This commit adds request specs for domain management. It usages
the `.all` interface to fetch the list of the domains and then it
usage the second domain to test the `#activate` and `#deactivate`
interface and finally it usages `.fetch` to retrieve the updated
details for that domain. So as long as this request specs passes
means this endpoint is working as expected.