Add request specs for certificate generation
This commit adds a request specs, that goes through the complete
order process and wait a certain time for the certificate to be
issued. One then download the certificate content and save it to
a temporary variable and finally revoke just created certificate

This also sets the `debug_mode` mode to `on` for actual API calls
as those are expensive and it's always handy to see what's going
on under the hood.

Note: We wait 10 seconds for the certificate to be approved, but
still there is a possibilities that it might fail, in that case
please check log or the Digicert account for details.