Remove coupling in the order's specs
Some order's specs were explicitly testing the outgoing messages from
the order class, but those interfaces has already been been tested in
their own class. These creates coupling in the specs, as if there are
any changes on the interface then we have to rewrite the order specs,
which is just a duplication and should not be necessary.

This commit decouple these specs and only verify that order class is
sending the outgoing messages properly without explicitly testing
their internal behavior.