Dynamically generate CSR using OpenSSL
This commit adds a CSR generator, which expects two hash keys and
it will generate the CSR content that can be used to request any
certificate through Digicert. We separated it as separate support
helper so we can use it from any of the specs file. Usage

```ruby
Digicert::CSRGenerator.generate(
domain: domain_name, organization: organization,
)
```