Add a configurable application debugger
We need to debug API request and Responses for different purpose,
This commit adds a debugger that will log requests and responses
details in the terminal. We have a added a config variable that
can be use to turn the debug mode on or off, by default it will
be off but we can turn that on by simply adding the `debug_mode`
configuration in the initializer or any other file.

```ruby
Digicert.configure do |digicert_config|
  digicert_config.debug_mode = true
end
```