Fix `#initialize_http_header` related issue
In the `Digicert::Request` we're using `#initialize_http_header` to
set Dev API key and `Content-Type` header separately. But the way
`#initialize_http_header`  every call to this method reset the
header and only sets the provided has key values pair.

This was causing our `create` request to fail, as for the create
request we need to specify the `content_type` with the `API Key`,
but it was being overridden in the way it was.

This commit fixes this, Now all of our headers should have been
setted up properly.

Source: https://github.com/ruby/ruby/blob/trunk/lib/net/http/header.rb#L14