Add request spec for domain creation
This commit rewrite the domain management's request specs, now
that we have the `create` operation working, so this commit is
creating a new domain and then playing with that domain to test
out different interfaces.

This commit also updates the `attributes` building for domain
creation as Digicert throws an error if the user passes upper
case `validation_type`. So let our client make sure no matter
what user's provides, we will `downcase` them before sending
it to the Digicert API.