Add certificate reissuing request spec
This commit adds the request specs to make sure the certificate
reissuing is working on the Actual API. This also expects us to
have a non expired certificate in our test environment and then
it will try to reissue that certificate.

In case this fails, then run the certificate generation request
spec first and then try this again, as certificate generation
will create an actual certificate that this spec can use.