Add Request Specs for Certificate Duplication
This commit adds the certificate duplication's request specs, this
also assume the last order to be ssl wildcard and it tries to use
that for the duplication. So, if this fails then run the order ssl
wildcard request specs, which will create an order and then re run
this spec should work fine.

In the underneath, this specs test the certificate duplication,
find duplicate certificate and as well the duplication listing
interfaces.

```sh
bin/rspec --tag api_call spec/requests/certificate_duplication_spec.rb
```