Add `san_names` support for CSR generation
This commit adds support for `san_names` in the CSR generation, it
also adds `r509` as development dependency to minimize the ground
work to use `openssl` directly.

Now we don't need to worry about sanitization or any other ground
work that we were doing to generate a CSR using `openssl`, and it
also seems pretty flexible so it would be easier to manage if we
need to extend any of it's functionality.

Ref: http://www.rubydoc.info/gems/r509/