Expose test stubs and helpers to developer
Actual API requests are slow and expensive and we try not to make
actual API request when possible. For most of our tests we mock the
API call which verifies the endpoint, http verb and headers, based
on those we helpers responses with an identical fixture file.

We have been using it internally and now we want developer to have
the option so they can test their application while using our gem.

This commit add a `digicert/rspec` file, so developer can require
this one in their `spec_helper` to utilize all of our test helper