Add support for query params in orders
Digicert supports to pass query params in the listing interfaces,
we already had that one implemented in the `all` action, but it
was not documented.

This commit adds the necessary test cases and document this to
the readme, we now can use this to retrieve filtered orders.

```ruby
Digicert::Order.all(limit: 20, offset: 0, sort: 'date_created')
```