Fix reissue/duplication issue for custom attribute
Normally the Order `reissue` and `duplication` interface takes an order
id and then usages the details from that order, and additionally it also
supports for custom attributes. The way it supposed to work is if we add
any custom attributes then it will use that along with all others but if
we don't then it will use the defaults.

It is working fine when no attributes is provided, but when we provide
custom attributes then it is replacing all of the attributes with the
new one, which might not include all the attributes, that's why we are
having issue like https://github.com/riboseinc/digicert-cli/issues/46

This commit fix that issue, and it also updates the existing interface
to provide the custom attributes along with the `order_id`, and it will
do the work of organizing those, bit less work for users!