Fix duplicate certificate finder bug
To find a duplicate certificate, we are using retrieving details
for any requests and then use the created_at date for the request
and order to find out the duplicate order.

But for some reason, looks like there might be couple of seconds
or differences between those dates, so this commit changes our
finder to look for any certificate with 5 seconds time and this
should fix the issue for now.

In the future, we should look for more stable solution, and if
necessary then we can also adjust the time.

Fixes #137