Fix empty duplicate certificate issue
When there are no duplicate certificate for any request, then the
certificate finder throws an error. But it might be more useful to
return nil if there are duplicate certificate for a request.

This commit takes that into consideration, and return nil when
there no duplicate certificate for any of the request.