Fix duplicate certificate finder issue
Recently, we came across an issue in the duplicate certificate
finder. Let's say one certificate has an existing duplicate and
in a later date we also request another one which is pending.
Now if we use the finder to find the duplicate certificate then
it would return the old one, which is miss leading.

This commit fixes this by checking the status of the request
and if it's still pending then it throws an error.

Discussion: 138