Fix SSL EV Plus duplication issue
Currently, the duplication for SSL EV Plus request throws an
`Invalid DNS name` issue. After looking into this we found this
interface does not expect us to provide both `root` and `www` in
the same request, otherwise it won't process that request.

This commit adds simple fix, and it will remove `www.` from
the `dns_names` list and this should fix the issue for now.

Fixes: 135